requirements_specific.txt
tqdm
omegaconf
iopath
einops
SentencePiece
opencv-python
ftfy
jupyterlab
ipdb

timm==0.6.13
decord==0.6.0
bitsandbytes==0.37.0
gradio==3.24.1
gradio-client==0.0.8
pytorchvideo==0.1.5
webdataset==0.2.48
omegaconf==2.3.0
transformers==4.28.0
accelerate==0.16.0